refactor(cli): share middleware request projection
